pstalgo: Fix for bug in SphereTree
pstalgo: Fix for bug in that caused network lines to not connect propertly, but only under very rare conditions. The problem only occurred when line endpoints meet at the exact centre of the map, and at the exact centre of each quadrant of the map, and at the exact centre of each quadrant of each quadrant, and so on for a couple of levels depending on the size of the map (at exact centre of each node in SphereTree).